refactor(engine): move files to renderer plugin dir
