+++ engine/src/cubos/engine/renderer/pps/pass.cpp
@@ -1,4 +1,4 @@
-#include <cubos/engine/gl/pps/pass.hpp>
+#include <cubos/engine/renderer/pps/pass.hpp>
 
 cubos::engine::gl::pps::Pass::Pass(core::gl::RenderDevice& renderDevice)
     : mRenderDevice(renderDevice)
